pyproject.toml
[tool.poetry]
name = "encost-test"
version = "0.1.0"
description = ""
authors = ["Alex Pavlov <yukiu217@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = {extras = ["classic"], version = "^0.30.6"}
fastapi = "^0.115.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.5.2"
psycopg2-binary = "^2.9.9"
tortoise-orm = {extras = ["aiosqlite"], version = "^0.21.6"}
loguru = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
